refactor: move BlockHashOrNumber to primitives
