+++ crates/primitives/src/lib.rs
@@ -29,7 +29,7 @@ mod transaction;
 pub mod proofs;
 
 pub use account::Account;
-pub use block::{Block, BlockLocked};
+pub use block::{Block, BlockHashOrNumber, BlockLocked};
 pub use chain::Chain;
 pub use constants::MAINNET_GENESIS;
 pub use forkid::{ForkFilter, ForkHash, ForkId, ValidationError};
